Optimizes prebuild script for faster execution
Improves the prebuild script by adding checks to avoid unnecessary image processing.
It now compares the hash of the source data and assets directory to the cached values.
If no changes are detected, the script skips the image processing step, resulting in faster execution times.

Also, increases webp quality to 100.